Add sync migrations.
